Santiago_FE

#' A dataset containing information about Feelings and emotions of users of active and motorized modes of transportation in Santiago sourced from 2016 survey.

data(Santiago_FE)

Format

An object of class tbl_df (inherits from tbl, data.frame) with 451 rows and 22 columns.

Source

A face to face survey considering pre-census 2012

Details

@format A data frame with 451 rows and 130 variables:

r2AA_FREEDOM

indicatation of the modes of transport that respondents relate to the following feelings and concepts: Freedom

r2AB_UNSAFETY

indicatation of the modes of transport that respondents relate to the following feelings and concepts: Unsafety

r2AC_FUNCTIONALITY

indicatation of of the modes of transport that respondents relate to the following feelings and concepts: Functionality

r2AD_ENJOYMENT

indicatation of the modes of transport that respondents relate to the following feelings and concepts: Enjoyment

r2AE_LOWCOST

indicatation of the modes of transport that respondents relate to the following feelings and concepts: Low cost

r2AF_POVERTY

indicatation of the modes of transport that respondents relate to the following feelings and concepts: Poverty

r2AG_SAFETY

indicatation of the modes of transport that respondents relate to the following feelings and concepts: Safety

r2AH_WASTE_OF_TIME

indicatation of the modes of transport that respondents relate to the following feelings and concepts: Waste of time

r2AI_UNPUNCTUALITY

indicatation of the modes of transport that respondents relate to the following feelings and concepts: Unpunctuality

r2AJ_CONGEST

indicatation of the modes of transport that respondents relate to the following feelings and concepts: Congestion

r2AK_EFFICIENCY

indicatation of the modes of transport that respondents relate to the following feelings and concepts: Efficiency

r2AL_LUXURY

indicatation of the modes of transport that respondents relate to the following feelings and concepts: Luxury

r2AM_ENVIRONMENT

indicatation of the modes of transport that respondents relate to the following feelings and concepts: Environmental care

r2AN_HEALTH

indicatation of the modes of transport that respondents relate to the following feelings and concepts: Health

r2AO_INTSOCI

indicatation of the modes of transport that respondents relate to the following feelings and concepts: Social Interaction

r2AP_UNCOMFT

indicatation of the modes of transport that respondents relate to the following feelings and concepts: Uncomfort

r2AQ_HAPPINESS

indicatation of the modes of transport that respondents relate to the following feelings and concepts: Happiness

r2AR_STATUS

indicatation of the modes of transport that respondents relate to the following feelings and concepts: Status

r2B_DAILY_ENJOY

How much do respondents usually enjoy traveling to their daily activities?

r2C_IMP_QUALITY

What level of importance do respondents attribute to the quality of their trips?

r2D_AFFECT

What does affect the most to their travel experience?

r2E_FACILIT

What does help respondents the most to their travel experience?

Examples

 data(Santiago_FE)
 r2AA_FREEDOM <- Santiago_FE$r2AA_FREEDOM