Santiago_RPD

#' A dataset containing information about reason and planning decision of users of active and motorized modes of transportation in Santiago sourced from 2016 survey.

data(Santiago_RPD)

Format

An object of class tbl_df (inherits from tbl, data.frame) with 451 rows and 23 columns.

Source

A face to face survey considering pre-census 2012

Details

@format A data frame with 451 rows and 130 variables:

r3A_ACCESS

How do respondents assess their access to employment opportunities through public transport?

r3B_ACC_EM

How does their access to public transport allows you to access the employment you need?

r3CA_FAM

How often do respondents perform these activities? Visiting family and friends

r3CB_REC

How often do respondents perform these activities? Recreational activities

r3CC_CUL

How often do respondents perform these activities? Cultural activities

r3CD_SPO

How often do respondents perform these activities? Sport activities

r3CE_GROC

How often do respondents perform these activities? Grocery/Food shopping

r3CF_SOC

How often do respondents perform these activities? Social activities

r3DA_FAM

NEEDS DESCRIPTION

r3DB_REC

NEEDS DESCRIPTION

r3DC_CUL

NEEDS DESCRIPTION

r3DD_SPO

NEEDS DESCRIPTION

r3DE_GROC

NEEDS DESCRIPTION

r3DF_SOC

NEEDS DESCRIPTION

r3E_OPTIONS

How important do respondents consider having several options to use different modes of transport?

r3F_ACCESS_DEPENDENCY

How much does respondents' quality of life depend on the access respondents currently have to public transport?

r3G_QUALITY_INCRS

How much do respondents think their quality of life would increase if they have better access to public transport?

r3H_ECON

Which transport modes are affordable for respondents and their family?

r3I_NOECON

Which transport modes are economically unaffordable?

r3JA_OFIC

How important it is to improve the following aspects of the public transport system? Improve access to offices and comerccial areas

r3JB_MODES

How important it is to improve the following aspects of the public transport system? Improve disponibility of different transport modes

r3JC_COMFORT

How important it is to improve the following aspects of the public transport system? Improve comfort for the use of public transport

r3JD_OTHERS

How important it is to improve the following aspects of the public transport system? Incorporate other modes to the fare system

Examples

 data(Santiago_RPD)
 r3A_ACCESS <- Santiago_RPD$r3A_ACCESS